tambah episode dan authentication login serta authorize

+++ app/controllers/episodes_controller.rb
@@ -1,10 +1,13 @@
 class EpisodesController < ApplicationController
+  before_action(:authenticate_user!, except: %i(index show))
   before_action(:set_episode, only: %i(show edit update destroy))
+
   def index
     @episodes = Episode.all
   end
 
   def new
+    authorize(Episode)
     @episode = Episodes::Create.new()
   end
 
@@ -12,23 +15,27 @@ class EpisodesController < ApplicationController
     success = -> (episode) { redirect_to(episode_path(episode), notice: 'success') }
     failure = -> (episode) { @episode = episode; render(:new); puts episode.errors.full_messages }
 
-    EpisodeService::Create.(episode_params, success: success, failure: failure)
+    authorize(Episode)
+    EpisodeService::Create.(current_user, episode_params, success: success, failure: failure)
   end
 
   def update
     success = -> (episode) { redirect_to(episode_path(episode), notice: 'success') }
     failure = -> (episode) { @episode = episode; render(:edit); puts episode.errors.full_messages }
 
-    EpisodeService::Update.(@episode, episode_params, success: success, failure: failure)
+    authorize(@episode)
+    EpisodeService::Update.(current_user, @episode, episode_params, success: success, failure: failure)
   end
 
   def show
   end
 
   def edit
+    authorize(@episode)
   end
 
   def destroy
+    authorize(@episode)
     redirect_to(episodes_path) if @episode.destroy
   end
